+++ src/queries/useAuthoritiesBrowse/useAuthoritiesBrowse.test.js
@@ -27,15 +27,20 @@ const wrapper = ({ children }) => (
   </Harness>
 );
 
+const decodeURIWithEquals = (string) => {
+  return decodeURI(string).replace(/%3D/g, '=');
+};
+
 describe('Given useAuthoritiesBrowse', () => {
   const filters = {};
   const searchQuery = 'test';
   const searchIndex = searchableIndexesValues.PERSONAL_NAME;
   const precedingRecordsCount = 5;
 
-  const generateTestAuthorities = (count, headingRef = 'authority') => new Array(count).fill({}).map((_, index) => ({
-    headingRef: `${headingRef}_${index}`,
-  }));
+  const generateTestAuthorities = (count, headingRef = 'authority') => new Array(count).fill({})
+    .map((_, index) => ({
+      headingRef: `${headingRef}_${index}`,
+    }));
 
   const mockGet = jest.fn();
 
@@ -71,15 +76,12 @@ describe('Given useAuthoritiesBrowse', () => {
     await waitFor(() => !result.current.isLoading);
 
     expect(mockGet).toHaveBeenCalledTimes(3);
-    expect(mockGet.mock.calls[0][0])
-      // browse/authorities?limit=20&precedingRecordsCount=5&query=(headingRef>="test" or headingRef<"test") and headingType==("Personal Name")
-      .toBe('browse/authorities?limit=20&precedingRecordsCount=5&query=%28headingRef%3E%3D%22test%22%20or%20headingRef%3C%22test%22%29%20and%20headingType%3D%3D%28%22Personal%20Name%22%29');
-    expect(mockGet.mock.calls[1][0])
-      // browse/authorities?limit=20&precedingRecordsCount=5&query=headingRef<"authority_0_0" and headingType==("Personal Name")
-      .toBe('browse/authorities?limit=20&precedingRecordsCount=5&query=headingRef%3C%22authority_0_0%22%20and%20headingType%3D%3D%28%22Personal%20Name%22%29');
-    expect(mockGet.mock.calls[2][0])
-      // browse/authorities?limit=20&precedingRecordsCount=5&query=headingRef>"authority_0_49" and headingType==("Personal Name")
-      .toBe('browse/authorities?limit=20&precedingRecordsCount=5&query=headingRef%3E%22authority_0_49%22%20and%20headingType%3D%3D%28%22Personal%20Name%22%29');
+    expect(decodeURIWithEquals(mockGet.mock.calls[0][0]))
+      .toBe('browse/authorities?limit=20&precedingRecordsCount=5&query=(headingRef>="test" or headingRef<"test") and isTitleHeadingRef==false and headingType==("Personal Name")');
+    expect(decodeURIWithEquals(mockGet.mock.calls[1][0]))
+      .toBe('browse/authorities?limit=20&precedingRecordsCount=5&query=headingRef<"authority_0_0" and isTitleHeadingRef==false and headingType==("Personal Name")');
+    expect(decodeURIWithEquals(mockGet.mock.calls[2][0]))
+      .toBe('browse/authorities?limit=20&precedingRecordsCount=5&query=headingRef>"authority_0_49" and isTitleHeadingRef==false and headingType==("Personal Name")');
   });
 
   describe('when requesting a previous page', () => {
@@ -105,9 +107,8 @@ describe('Given useAuthoritiesBrowse', () => {
       await waitFor(() => !result.current.isLoading);
 
       expect(mockGet).toHaveBeenCalledTimes(1);
-      expect(mockGet.mock.calls[0][0])
-        // browse/authorities?limit=20&precedingRecordsCount=5&query=headingRef<"authority_1_0" and headingType==("Personal Name")
-        .toBe('browse/authorities?limit=20&precedingRecordsCount=5&query=headingRef%3C%22authority_1_0%22%20and%20headingType%3D%3D%28%22Personal%20Name%22%29');
+      expect(decodeURIWithEquals(mockGet.mock.calls[0][0]))
+        .toBe('browse/authorities?limit=20&precedingRecordsCount=5&query=headingRef<"authority_1_0" and isTitleHeadingRef==false and headingType==("Personal Name")');
     });
   });
 
@@ -173,15 +174,12 @@ describe('Given useAuthoritiesBrowse', () => {
       await waitFor(() => !result.current.isLoading);
 
       expect(mockGet).toHaveBeenCalledTimes(6);
-      expect(mockGet.mock.calls[3][0])
-        // browse/authorities?limit=20&precedingRecordsCount=5&query=(headingRef>="test2" or headingRef<"test2") and headingType==("Personal Name")
-        .toBe('browse/authorities?limit=20&precedingRecordsCount=5&query=%28headingRef%3E%3D%22test2%22%20or%20headingRef%3C%22test2%22%29%20and%20headingType%3D%3D%28%22Personal%20Name%22%29');
-      expect(mockGet.mock.calls[4][0])
-        // browse/authorities?limit=20&precedingRecordsCount=5&query=headingRef<"authority_3_0" and headingType==("Personal Name")
-        .toBe('browse/authorities?limit=20&precedingRecordsCount=5&query=headingRef%3C%22authority_3_0%22%20and%20headingType%3D%3D%28%22Personal%20Name%22%29');
-      expect(mockGet.mock.calls[5][0])
-        // browse/authorities?limit=20&precedingRecordsCount=5&query=headingRef>"authority_3_49" and headingType==("Personal Name")
-        .toBe('browse/authorities?limit=20&precedingRecordsCount=5&query=headingRef%3E%22authority_3_49%22%20and%20headingType%3D%3D%28%22Personal%20Name%22%29');
+      expect(decodeURIWithEquals(mockGet.mock.calls[3][0]))
+        .toBe('browse/authorities?limit=20&precedingRecordsCount=5&query=(headingRef>="test2" or headingRef<"test2") and isTitleHeadingRef==false and headingType==("Personal Name")');
+      expect(decodeURIWithEquals(mockGet.mock.calls[4][0]))
+        .toBe('browse/authorities?limit=20&precedingRecordsCount=5&query=headingRef<"authority_3_0" and isTitleHeadingRef==false and headingType==("Personal Name")');
+      expect(decodeURIWithEquals(mockGet.mock.calls[5][0]))
+        .toBe('browse/authorities?limit=20&precedingRecordsCount=5&query=headingRef>"authority_3_49" and isTitleHeadingRef==false and headingType==("Personal Name")');
     });
   });